Conflicts:
	projects/core/core.vcxproj.filters
	src/graphics/renderer.cpp

+++ src/core/json_serializer.h
@@ -61,7 +61,7 @@ namespace Lux
 			inline void writeString(const char* str)
 			{
 				m_file.write("\"", 1);
-				m_file.write(str, strlen(str));
+				m_file.write(str, (int32_t)strlen(str));
 				m_file.write("\"", 1);
 			}
