+++ src/core/memory_tracker.h
@@ -35,14 +35,14 @@ namespace Lux
 				: m_file(src.m_file), m_line(src.m_line), m_alloc_id(src.m_alloc_id), m_size(src.m_size), m_mark(src.m_mark)
 			{}
 
-			Entry(const char* file, int line, size_t size)
+			Entry(const char* file, int line, intptr_t size)
 				: m_file(file), m_line(line), m_alloc_id(MemoryTracker::getAllocID()), m_size(size), m_mark(0)
 			{}
 
 			const char* file() const { return m_file; }
 			const int line() const { return m_line; }
 			const uint32_t allocID() const { return m_alloc_id; }
-			const size_t size() const { return m_size; }
+			const intptr_t size() const { return m_size; }
 
 			void mark() { ++m_mark; }
 			const uint8_t getMark() const { return m_mark; }
@@ -51,7 +51,7 @@ namespace Lux
 			const char* m_file;
 			uint32_t m_line;
 			uint32_t m_alloc_id;
-			size_t m_size;
+			intptr_t m_size;
 			uint8_t m_mark;
 		};
 
@@ -59,7 +59,7 @@ namespace Lux
 		static MemoryTracker& getInstance();
 		static void destruct();
 
-		void add(void* p, const size_t size, const char* file, const int line);
+		void add(void* p, const intptr_t size, const char* file, const int line);
 		void remove(void* p);
 
 		void dumpDetailed();
@@ -83,7 +83,7 @@ namespace Lux
 		EntryTable m_map;
 
 		MT::SpinMutex m_spin_mutex;
-		uint32_t m_allocated_memory;
+		intptr_t m_allocated_memory;
 		uint8_t m_mark;
 
 		static MemoryTracker* s_instance;
